backend/pyproject.toml
[project]
name = "party-registration-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi[standard]~=0.128.0",
  "uvicorn~=0.40.0",
  "sqlalchemy[asyncio]~=2.0.0",
  "pydantic-settings~=2.12.0",
  "asyncpg~=0.31.0",
  "psycopg2-binary~=2.9.0",
  "bcrypt~=5.0.0",
  "pytest~=9.0.0",
  "pytest-asyncio~=1.3.0",
  "aiosqlite~=0.22.0",
  "googlemaps~=4.10.0",
  "pyjwt~=2.10.0",
]

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["test"]
python_files = "*_test.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
addopts = "--color=yes"
